[#165407760] Rename files according to engineering-book
