+++ server/models/vote.model.js
