--- server/routes/voteRoutes.js
@@ -1,9 +1,0 @@
-import express from 'express';
-
-import saveVote from '../controllers/voteController.js';
-
-const router = express.Router();
-
-router.post('/', saveVote);
-
-export default router;
